+++ braindecode/models/eegconformer.py
@@ -10,6 +10,7 @@ from einops.layers.torch import Rearrange
 from torch import nn, Tensor
 
 from braindecode.models.base import EEGModuleMixin
+from braindecode.modules import MultiHeadAttention, FeedForwardBlock
 
 
 class EEGConformer(EEGModuleMixin, nn.Module):
@@ -245,35 +246,6 @@ class _PatchEmbedding(nn.Module):
         return x
 
 
-class _MultiHeadAttention(nn.Module):
-    def __init__(self, emb_size, num_heads, dropout):
-        super().__init__()
-        self.emb_size = emb_size
-        self.num_heads = num_heads
-        self.keys = nn.Linear(emb_size, emb_size)
-        self.queries = nn.Linear(emb_size, emb_size)
-        self.values = nn.Linear(emb_size, emb_size)
-        self.att_drop = nn.Dropout(dropout)
-        self.projection = nn.Linear(emb_size, emb_size)
-
-    def forward(self, x: Tensor, mask: Tensor = None) -> Tensor:
-        queries = rearrange(self.queries(x), "b n (h d) -> b h n d", h=self.num_heads)
-        keys = rearrange(self.keys(x), "b n (h d) -> b h n d", h=self.num_heads)
-        values = rearrange(self.values(x), "b n (h d) -> b h n d", h=self.num_heads)
-        energy = torch.einsum("bhqd, bhkd -> bhqk", queries, keys)
-        if mask is not None:
-            fill_value = torch.finfo(torch.float32).min
-            energy.mask_fill(~mask, fill_value)
-
-        scaling = self.emb_size ** (1 / 2)
-        att = F.softmax(energy / scaling, dim=-1)
-        att = self.att_drop(att)
-        out = torch.einsum("bhal, bhlv -> bhav ", att, values)
-        out = rearrange(out, "b h n d -> b n (h d)")
-        out = self.projection(out)
-        return out
-
-
 class _ResidualAdd(nn.Module):
     def __init__(self, fn):
         super().__init__()
@@ -284,16 +256,6 @@ class _ResidualAdd(nn.Module):
         x = self.fn(x, **kwargs)
         x += res
         return x
-
-
-class _FeedForwardBlock(nn.Sequential):
-    def __init__(self, emb_size, expansion, drop_p, activation: nn.Module = nn.GELU):
-        super().__init__(
-            nn.Linear(emb_size, expansion * emb_size),
-            activation(),
-            nn.Dropout(drop_p),
-            nn.Linear(expansion * emb_size, emb_size),
-        )
 
 
 class _TransformerEncoderBlock(nn.Sequential):
@@ -309,14 +271,14 @@ class _TransformerEncoderBlock(nn.Sequential):
             _ResidualAdd(
                 nn.Sequential(
                     nn.LayerNorm(emb_size),
-                    _MultiHeadAttention(emb_size, att_heads, att_drop),
+                    MultiHeadAttention(emb_size, att_heads, att_drop),
                     nn.Dropout(att_drop),
                 )
             ),
             _ResidualAdd(
                 nn.Sequential(
                     nn.LayerNorm(emb_size),
-                    _FeedForwardBlock(
+                    FeedForwardBlock(
                         emb_size,
                         expansion=forward_expansion,
                         drop_p=att_drop,
